Adicionando sistema de adicionar e devolver

+++ src/main/java/br/edu/fesa/biblioteca/controller/UsuarioController.java
@@ -29,20 +29,20 @@ import org.springframework.web.multipart.MultipartFile;
 @Controller
 @RequestMapping("/Usuario")
 public class UsuarioController {
-
+    
     @Autowired
     private UsuarioService usuarioService;
-
+    
     @Autowired
     private UsuarioRepository usuarioRepository;
-
+    
     @Autowired
     private TokenService tokenService;
 
     // Exibe a página de cadastro
     @GetMapping("/cadastro")
     public String mostrarFormularioCadastro(@RequestParam(name = "sucesso", required = false) String sucesso, Model model, HttpServletRequest request) throws UnsupportedEncodingException {
-
+        
         if (sucesso != null) {
             model.addAttribute("sucesso", true);
         } else {
@@ -51,21 +51,21 @@ public class UsuarioController {
         model.addAttribute("usuario", new Usuario());
         return "Usuario/cadastro";
     }
-
+    
     @PostMapping("/cadastrar")
     public String cadastrarUsuario(@ModelAttribute Usuario usuario, Model model, @RequestParam("arquivoImagem") MultipartFile arquivoImagem,
             RedirectAttributes redirectAttributes, HttpServletRequest request) throws UnsupportedEncodingException {
         try {
             Usuario usuarioNovo = new Usuario();
             ModelMap erros = verificaerros(usuario);
-
+            
             if (!erros.isEmpty()) {
                 // Copia os erros para o model principal
                 model.addAllAttributes(erros);
                 model.addAttribute("usuario", usuario); // opcional: manter os dados preenchidos
                 return "Usuario/cadastro";
             }
-
+            
             if (!arquivoImagem.isEmpty()) {
                 String nomeArquivo = arquivoImagem.getOriginalFilename();
                 usuario.setImagem(arquivoImagem.getBytes());
@@ -87,22 +87,22 @@ public class UsuarioController {
             return "Usuario/cadastro"; // Volta para a página de cadastro com os dados já preenchidos
         }
     }
-
+    
     @GetMapping("/lista")
     public String listar(ModelMap model, HttpServletRequest request) throws UnsupportedEncodingException {
         List<Usuario> Usuarios = usuarioService.findAll();
-
+        
         List<Usuario> sortedUsuarios = Usuarios.stream()
                 .sorted((padrao1, padrao2) -> padrao1.getNome().compareTo(padrao2.getNome()))
                 .collect(Collectors.toList());
         model.addAttribute("usuarios", sortedUsuarios);
         return "Usuario/listarUsuario";
     }
-
+    
     @GetMapping("/Meu_usuario")
     public String meuUsuario(HttpServletRequest request, RedirectAttributes redirectAttributes, Model model) throws UnsupportedEncodingException {
         try {
-
+            
             String usuarioId = (String) request.getSession(false).getAttribute("usuarioId");
             UUID uuid = UUID.fromString(usuarioId);
             Usuario user = usuarioService.findById(uuid).get();
@@ -115,7 +115,7 @@ public class UsuarioController {
             return "redirect:/biblioteca-fesa";
         }
     }
-
+    
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @GetMapping("/editar/{id}")
     public String editar(@PathVariable UUID id, ModelMap model, HttpServletRequest request) throws UnsupportedEncodingException {
@@ -125,12 +125,24 @@ public class UsuarioController {
         model.addAttribute("origem", "/Usuario/lista");
         return "Usuario/editarUsuario";
     }
-
+    
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @PostMapping("/atualizar/{id}")
     public String atualizar(@PathVariable UUID id, @Valid
             @ModelAttribute Usuario usuario, BindingResult bindingResult, ModelMap model, HttpServletRequest request, @RequestParam("arquivoImagem") MultipartFile arquivoImagem) throws UnsupportedEncodingException, IOException {
+        
+        int numeroAdmin;
+        
         Usuario user = usuarioService.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));;
+        numeroAdmin = (int) usuarioRepository.countByAdminTrue();
+        
+        if (user.isAdmin() && numeroAdmin == 1 && !usuario.isAdmin()) {
+            user.setImagemEmbase64();
+            model.addAttribute("usuario", user);
+            model.addAttribute("erro", "É necessário no minimo um usuário administrador no sistema");
+            return "Usuario/editarUsuario";
+        }
+        
         if (!arquivoImagem.isEmpty()) {
             String nomeArquivo = arquivoImagem.getOriginalFilename();
             user.setImagem(arquivoImagem.getBytes());
@@ -138,13 +150,14 @@ public class UsuarioController {
         }
         if (bindingResult.hasErrors()) {
             model.addAttribute("usuario", usuario);
-            return "/usuario/editar";
+            return "Usuario/editarUsuario";
         }
-
+        
         user.setEmail(usuario.getEmail());
         user.setNome(usuario.getNome());
         user.setSenha(usuario.getSenha());
         user.setTelefone(usuario.getTelefone());
+        user.setAdmin(usuario.isAdmin());
         usuarioService.update(user);
         if (user.isAdmin()) {
             return "redirect:/Usuario/lista";
@@ -152,31 +165,42 @@ public class UsuarioController {
             return "redirect:/home";
         }
     }
-
+    
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @GetMapping("/salvar")
     public String salvar(Usuario usuario, HttpServletRequest request, Model model) throws UnsupportedEncodingException {
         usuarioService.save(usuario);
         return "Usuario/listarUsuario";
     }
-
+    
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @PostMapping("/excluir/{id}")
     public String confirmarExclusao(@PathVariable UUID id, Model model, HttpServletRequest request) throws UnsupportedEncodingException {
+        int numeroAdmin;
+        
+        Usuario user = usuarioService.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));;
+        numeroAdmin = (int) usuarioRepository.countByAdminTrue();
+        
+        if (user.isAdmin() && numeroAdmin == 1) {
+            model.addAttribute("usuario", user);
+            model.addAttribute("erro", "É necessário no minimo um usuário administrador no sistema");
+            return "Usuario/remover";
+        }
+         
         usuarioService.deleteById(id);
         return "redirect:/Usuario/lista";
     }
-
+    
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @GetMapping("/remover/{id}")
     public String remover(@PathVariable UUID id, Model model, HttpServletRequest request) throws UnsupportedEncodingException {
         model.addAttribute("usuario", usuarioService.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado")));
         return "Usuario/remover";
     }
-
+    
     public ModelMap verificaerros(Usuario usuario) {
         ModelMap erros = new ModelMap();
-
+        
         if (usuario.getNome().length() < 10) {
             erros.addAttribute("ErrorNome", "Favor insira o nome completo");
         }
